+++ index.js
@@ -21,8 +21,8 @@ const fastifyEnv = require('@fastify/env')
 const fastifyMultipart = require('@fastify/multipart')
 
 const ajvFormats = require('ajv-formats')
-const lunset = require('lodash.unset')
 
+const { unset: lunset } = require('lodash')
 const { readdirSync } = require('fs')
 const { join } = require('path')
 const { ObjectId } = require('mongodb')
@@ -47,6 +47,7 @@ const {
 const { registerMongoInstances } = require('./lib/mongo/mongo-plugin')
 const { getAjvResponseValidationFunction } = require('./lib/validatorGetters')
 const { pointerSeparator } = require('./lib/JSONPath.utils')
+const { registerHelperRoutes } = require('./lib/helpersRoutes')
 
 async function registerCrud(fastify, { modelName, isView }) {
   if (!fastify.mongo) { throw new Error('`fastify.mongo` is undefined!') }
@@ -208,14 +209,20 @@ async function iterateOverCollectionDefinitionAndRegisterCruds(fastify) {
 const validCrudFolder = path => !['.', '..'].includes(path) && /\.js(on)?$/.test(path)
 
 async function setupCruds(fastify) {
-  const collections = readdirSync(fastify.config.COLLECTION_DEFINITION_FOLDER)
+  const {
+    COLLECTION_DEFINITION_FOLDER,
+    VIEWS_DEFINITION_FOLDER,
+    HELPERS_PREFIX,
+  } = fastify.config
+
+  const collections = readdirSync(COLLECTION_DEFINITION_FOLDER)
     .filter(validCrudFolder)
-    .map(path => join(fastify.config.COLLECTION_DEFINITION_FOLDER, path))
+    .map(path => join(COLLECTION_DEFINITION_FOLDER, path))
     .map(require)
 
   fastify.decorate('collections', collections)
 
-  const viewsFolder = fastify.config.VIEWS_DEFINITION_FOLDER
+  const viewsFolder = VIEWS_DEFINITION_FOLDER
   if (viewsFolder) {
     const views = readdirSync(viewsFolder)
       .filter(validCrudFolder)
@@ -231,6 +238,7 @@ async function setupCruds(fastify) {
       .register(fp(loadModels))
       .register(iterateOverCollectionDefinitionAndRegisterCruds)
       .register(joinPlugin, { prefix: '/join' })
+      .register(registerHelperRoutes, { prefix: HELPERS_PREFIX })
   }
 }