+++ tests/expectedSchemas/booksDeleteSchema.js
@@ -242,6 +242,15 @@ module.exports = {
             'items': {
               'type': 'object',
               'additionalProperties': true,
+              'properties': {
+                'edition': {
+                  'type': 'number',
+                },
+                'date': {
+                  'type': 'string',
+                  'format': 'date-time',
+                },
+              },
               'nullable': true,
             },
             'nullable': true,
@@ -249,6 +258,15 @@ module.exports = {
           {
             'type': 'object',
             'additionalProperties': true,
+            'properties': {
+              'edition': {
+                'type': 'number',
+              },
+              'date': {
+                'type': 'string',
+                'format': 'date-time',
+              },
+            },
             'nullable': true,
           },
         ],
@@ -329,6 +347,14 @@ module.exports = {
       },
       'attachments\\.\\d+\\.more\\.\\d+$': {
         'type': 'string',
+      },
+      'editionsDates\\.\\d+\\..+$': true,
+      'editionsDates\\.\\d+\\.edition$': {
+        'type': 'number',
+      },
+      'editionsDates\\.\\d+\\.date$': {
+        'type': 'string',
+        'format': 'date-time',
       },
     },
     'additionalProperties': false,