+++ tests/expectedSchemas/booksExportSchema.js
@@ -249,6 +249,15 @@ module.exports = {
             'items': {
               'type': 'object',
               'additionalProperties': true,
+              'properties': {
+                'edition': {
+                  'type': 'number',
+                },
+                'date': {
+                  'type': 'string',
+                  'format': 'date-time',
+                },
+              },
               'nullable': true,
             },
             'nullable': true,
@@ -256,6 +265,15 @@ module.exports = {
           {
             'type': 'object',
             'additionalProperties': true,
+            'properties': {
+              'edition': {
+                'type': 'number',
+              },
+              'date': {
+                'type': 'string',
+                'format': 'date-time',
+              },
+            },
             'nullable': true,
           },
         ],
@@ -369,6 +387,14 @@ module.exports = {
       },
       'attachments\\.\\d+\\.more\\.\\d+$': {
         'type': 'string',
+      },
+      'editionsDates\\.\\d+\\..+$': true,
+      'editionsDates\\.\\d+\\.edition$': {
+        'type': 'number',
+      },
+      'editionsDates\\.\\d+\\.date$': {
+        'type': 'string',
+        'format': 'date-time',
       },
     },
     'additionalProperties': false,
@@ -683,6 +709,15 @@ module.exports = {
                 'items': {
                   'type': 'object',
                   'additionalProperties': true,
+                  'properties': {
+                    'edition': {
+                      'type': 'number',
+                    },
+                    'date': {
+                      'type': 'string',
+                      'format': 'date-time',
+                    },
+                  },
                   'nullable': true,
                 },
                 'nullable': true,
@@ -690,6 +725,15 @@ module.exports = {
               {
                 'type': 'object',
                 'additionalProperties': true,
+                'properties': {
+                  'edition': {
+                    'type': 'number',
+                  },
+                  'date': {
+                    'type': 'string',
+                    'format': 'date-time',
+                  },
+                },
                 'nullable': true,
               },
             ],