+++ tests/expectedSchemas/booksNewChangeStateManySchema.js
@@ -312,6 +312,15 @@ module.exports = {
                   'items': {
                     'type': 'object',
                     'additionalProperties': true,
+                    'properties': {
+                      'edition': {
+                        'type': 'number',
+                      },
+                      'date': {
+                        'type': 'string',
+                        'format': 'date-time',
+                      },
+                    },
                     'nullable': true,
                   },
                   'nullable': true,
@@ -319,6 +328,15 @@ module.exports = {
                 {
                   'type': 'object',
                   'additionalProperties': true,
+                  'properties': {
+                    'edition': {
+                      'type': 'number',
+                    },
+                    'date': {
+                      'type': 'string',
+                      'format': 'date-time',
+                    },
+                  },
                   'nullable': true,
                 },
               ],
@@ -513,6 +531,27 @@ module.exports = {
             },
             'attachments\\.\\d+\\.more\\.\\d+$': {
               'type': 'string',
+            },
+            'editionsDates\\.\\d+$': {
+              'type': 'object',
+              'properties': {
+                'edition': {
+                  'type': 'number',
+                },
+                'date': {
+                  'type': 'string',
+                  'format': 'date-time',
+                },
+              },
+              'additionalProperties': true,
+            },
+            'editionsDates\\.\\d+\\..+$': true,
+            'editionsDates\\.\\d+\\.edition$': {
+              'type': 'number',
+            },
+            'editionsDates\\.\\d+\\.date$': {
+              'type': 'string',
+              'format': 'date-time',
             },
           },
         },