+++ tests/httpInterface.getList.test.js
@@ -202,6 +202,12 @@ tap.test('HTTP GET /', async t => {
       found: HTTP_PUBLIC_FIXTURES.filter(f => f.price > 20 && !f.additionalInfo),
     },
     {
+      name: 'with stringified ISO Date in query filter',
+      url: `/?_q=${JSON.stringify({ 'editionsDates.date': { $lt: new Date('2020').toISOString() } })}`,
+      acl_rows: undefined,
+      found: HTTP_PUBLIC_FIXTURES.filter(f => Boolean(f.editionsDates)),
+    },
+    {
       name: 'with filter by additionalInfo nested with dot notation',
       url: `/?_q=${JSON.stringify({ 'additionalInfo.notes.mynote': 'good' })}`,
       acl_rows: undefined,
@@ -918,7 +924,7 @@ tap.test('HTTP GET / ', async t => {
 
   t.test('serialize correctly data on GET (a string should be casted to number to match schema)', async t => {
     const DOC = {
-      ...fixtures[0],
+      ...fixtures[1],
       // expected to be casted to number
       price: '44',
       ignoreMe: 'expect to be ignored',